+++ services/web/server/tests/integration/01/test_garbage_collection.py
@@ -48,7 +48,7 @@ from simcore_service_webserver.resource_manager.registry import (
     RegistryKeyPrefixDict,
     get_registry,
 )
-from simcore_service_webserver.rest.rest_plugin import setup_rest
+from simcore_service_webserver.rest.plugin import setup_rest
 from simcore_service_webserver.security.plugin import setup_security
 from simcore_service_webserver.session.plugin import setup_session
 from simcore_service_webserver.socketio.plugin import setup_socketio
