+++ services/web/server/tests/unit/with_dbs/03/garbage_collector/test_resource_manager.py
@@ -52,7 +52,7 @@ from simcore_service_webserver.resource_manager.registry import (
     RegistryKeyPrefixDict,
     get_registry,
 )
-from simcore_service_webserver.rest.rest_plugin import setup_rest
+from simcore_service_webserver.rest.plugin import setup_rest
 from simcore_service_webserver.security.plugin import setup_security
 from simcore_service_webserver.session.plugin import setup_session
 from simcore_service_webserver.socketio.messages import SOCKET_IO_PROJECT_UPDATED_EVENT
